\name{ex0116}
\alias{ex0116}
\docType{data}
\title{Gross Domestic Product (GDP) per Capita}
\description{
  The data are the gross domestic product per capita for 228 countries in 2010.
}
\usage{ex0116}
\format{
  A data frame with 228 observations on the following 3 variables.
  \describe{
    \item{\code{Rank}}{rank order of country from highest to lowest GDP}
    \item{\code{Country}}{name of country}
    \item{\code{PerCapitaGDP}}{per capita GDP in \$US}
  }
}
\source{
Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Central Intelligence Agency, Country Comparison: GDP per capita
  (PPP), \bold{The World Factbook},
  \url{https://www.cia.gov/library/publications/theworld-factbook/rankorder/2004rank.html}
  (retrieved June 30,2011).
  }
\examples{
str(ex0116)
}
\keyword{datasets}
