\name{ex0125}
\alias{ex0125}
\docType{data}
\title{Zinc concentrations for two groups of rats}
\description{
  The data are the zinc concentrations (in mg/ml) in the blood of rats that received a dietary supplement and rats that did not receive the supplement.
}
\usage{ex0125}
\format{
  A data frame with 39 observations on the following 39 variables.
  \describe{
    \item{\code{Group}}{a factor representing the group, with levels "A" for the dietary supplemnt group and "B" for the control group }
    \item{\code{Zinc}}{measured zinc concentration in mg/ml}
  }
}
\source{
Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}

\examples{
str(ex0125)
}
\keyword{datasets}
