\name{ex0330}
\alias{ex0330}
\docType{data}
\title{Education and Income}
\description{
  The data are incomes in U.S. dollars for 1,020 working Americans who had 12
  years of education and 406 working Americans who had 16 years of education,
  in 2005.
}
\usage{ex0330}
\format{
  A data frame with 1,426 observations on the following 3 variables.
  \describe{
    \item{\code{Subject}}{a subject identification number}
    \item{\code{Educ}}{number of years of education--either 12 or 16}
    \item{\code{Income2005}}{income, in dollars}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  National Longitudinal Survey of Youth, U.S. Bureau of Labor Statistics, 
  http://www.bls.gov/nls/home.htm (May 8, 2008). 
}
\seealso{
  \code{\link{ex0222}}, \code{\link{ex0331}}, \code{\link{ex0524}},
  \code{\link{ex0525}}, \code{\link{ex0828}}, \code{\link{ex0923}},
  \code{\link{ex1033}}, \code{\link{ex1223}}
}
\examples{
str(ex0330)
}
\keyword{datasets}
