\name{ex0430}
\alias{ex0430}
\docType{data}
\title{Sunlight Protection Factor}
\description{
  Tolerance to sunlight (in minutes) for 13 patients prior to and after
  treatment with a sunscreen. 
}
\usage{ex0430}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{\code{PreTreatment}}{tolerance to sunlight (minutes) prior to
      sunscreen application} 
    \item{\code{Sunscreen}}{tolerance to sunlight (minutes) after
      sunscreen application} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Fusaro, R.M. and Johnson, J.A. (1974).  Sunlight Protection for
  Erythropoietic Protoporphyria Patients, \emph{Journal of the American
  Medical Association} \bold{229}(11): 1420.
}
\examples{
str(ex0430)
}
\keyword{datasets}
