\name{ex0432}
\alias{ex0432}
\docType{data}
\title{Therapeutic Marijuana}
\description{
  To investigate the capacity of marijuana to reduce the side effects of
  cancer chemotherapy, researchers performed a double-blind, randomized,
  crossover trial.  Fifteen cancer patients on chemotherapy were
  randomly assigned to receive either a marijuana treatment or a placebo
  treatment after their first three sessions of chemotherapy.  They were
  then crossed over to the opposite treatment for their next 3 sessions.
}
\usage{ex0432}
\format{
  A data frame with 15 observations on the following 3 variables.
  \describe{
    \item{\code{Subject}}{subject number 1--15}
    \item{\code{Marijuana}}{total number of vomiting and retching
      episodes under marijuana treatment} 
    \item{\code{Placebo}}{total number of vomiting and retching episodes
      under placebo treatment} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Chang, A.E., Shiling, D.J., Stillman, R.C., Goldberg, N.H., Seipp,
  C.A., Barofsky, I., Simon, R.M. and Rosenberg,
  S.A. (1979). Delta-9-Tetrahydrocannabinol as an Antiemetic in Cancer
  Patients Receiving High Dose Methotrexate, \emph{Annals of Internal
  Medicine} \bold{91}(6): 819--824.
}
\examples{
str(ex0432)
}
\keyword{datasets}
