\name{ex0523}
\alias{ex0523}
\docType{data}
\title{Was Tyrannosaurus Rex Warm-Blooded?}
\description{
  Data frame with measurements of oxygen isotopic composition of
  vertebrate bone phosphate (per mil deviations from SMOW) in 12 bones
  of a singe Tyrannosaurus rex specimen
}
\usage{ex0523}
\format{
  A data frame with 52 observations on the following 2 variables.
  \describe{
    \item{\code{Oxygen}}{oxygen isotopic composition}
    \item{\code{Bone}}{a factor with levels \code{"Bone1"},
      \code{"Bone2"}, \dots, \code{"Bone12"}}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Barrick, R.E. and Showers, W.J. (1994).  Thermophysiology of
  \emph{Tyrannosaurus rex}: Evidence from Oxygen Isotopes,
  \emph{Science} \bold{265}(5169): 222--224.
}
\seealso{
  \code{\link{ex1120}}
}
\examples{
str(ex0523)
}
\keyword{datasets}
