\name{ex0722}
\alias{ex0722}
\docType{data}
\title{Crab Claw Size and Force}
\description{
  As part of a study of the effects of predatory intertidal crab species
  on snail populations, researchers measured the mean closing forces and
  the propdus heights of the claws on several crabs of three species.
}
\usage{ex0722}
\format{
  A data frame with 38 observations on the following 3 variables.
  \describe{
    \item{\code{Force}}{closing strength of claw of the crab}
    \item{\code{Height}}{propodus height of claw of the crab}
    \item{\code{Species}}{species to which the crab belongs}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Yamada, S.B. and Boulding, E.G. (1992). Claw Morphology, Prey Size Selection 
  and Foraging Efficiency in Generalist and Specialist Shell-Breaking Crabs.
  \emph{Journal of Experimental Marine Biology and Ecolog}, \bold{220} 191--211.
}
\examples{
str(ex0722)
}
\keyword{datasets}
