\name{ex0726}
\alias{ex0726}
\docType{data}
\title{Orign of the Term Regression}
\description{
  These data are heights of 933 adults and their parents, as measured by
  Karl Pearson in 1885.
}
\usage{ex0726}
\format{
  A data frame with 933 observations on the following 5 variables.
  \describe{
    \item{\code{Gender}}{a factor with levels \code{"female"} and \code{"male"}}
    \item{\code{Family}}{an identification number for family, 1,2,...,205  )}
    \item{\code{Height}}{adult height of the child, inches}
    \item{\code{Father}}{height of the child's father, inches )} 
    \item{\code{Mother}}{height of the child's mother, inches )}    
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Hubble, E. and Humason, M. (1931).  The Velocity--Distance Relation
  Among Extra--calactic Nebulae, \emph{Astrophysics Journal} \bold{74}:
  43--50. 
}
\seealso{
  \code{\link{ex1032}} and \code{\link{ex1618}}
}
\examples{
str(ex0725)
}
\keyword{datasets}
