\name{ex0824}
\alias{ex0824}
\docType{data}
\title{Respiratory Rates for Children}
\description{

  A high respiratory rate is a potential diagnostic indicator of
  respiratory infection in children.  To judge whether a respiratory
  rate is \dQuote{high} however, a physician must have a clear picture of the
  distribution of normal rates.  To this end, Italian researchers
  measured the respiratory rates of 618 children between the ages of 15
  days and 3 years. 
}
\usage{ex0824}
\format{
  A data frame with 618 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{age in months of child}
    \item{\code{Rate}}{respiratory rate (breaths per minute)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Rusconi, F., Castagneto, M., Porta, N., Gagliardi, L., Leo, G.,
  Pellegatta, A., Razon, S. and Braga, M. (1994).  Reference Values for
  Respiratory Rate in the First 3 Years of Life, \emph{Pediatrics}
  \bold{94}(3): 350--355. 
}
\examples{
str(ex0824)
}
\keyword{datasets}
