\name{ex0828}
\alias{ex0828}
\docType{data}
\title{IQ, Education, and Future Income}
\description{
  These data are armed Forces Qualifying Test (AFQT) score percentiles, years of
  education, and annual income in 2005 for a subset of a random sample of 2,584 
  Americans selected in 1979 who were working in 2005 and re-interviewed in 2006.
}
\usage{ex0828}
\format{
  A data frame with 2,584 observations on the following 2 variables.
  \describe{
    \item{\code{Subject}}{the subject identification number}
    \item{\code{AFQT}}{percentile score on the AFQT test}
    \item{\code{Educ}}{years of education achieved by 2005}
    \item{\code{Income2005}}{annual income in 2005}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  National Longitudinal Survey of Youth, U.S. Bureau of Labor Statistics, 
  http://www.bls.gov/nls/home.htm (May 8, 2008).
}
\seealso{
  \code{\link{ex0222}}, \code{\link{ex0330}}, \code{\link{ex0331}},
  \code{\link{ex0524}}, \code{\link{ex0525}}, \code{\link{ex0923}},
  \code{\link{ex1033}}, \code{\link{ex1223}}, \code{\link{ex1616}},
  \code{\link{ex1617}}, \code{\link{ex1717}}, \code{\link{ex1718}},
  \code{\link{ex2020}}
}
\examples{
str(ex0828)
}
\keyword{datasets}
