\name{ex0829}
\alias{ex0829}
\docType{data}
\title{Autism Rates}
\description{
  These data are the prevalence of autism per 10,000 ten-year old children
  in the United States in 1992, 1994, 1996, 1998, and 2000.
}
\usage{ex0829}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{Prevalence}}{the number of autism cases per 10,000 ten-year old 
      children}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
   Newschaffer, C. J.,  Falb, M. D. and Gurney, J. G. (2005) National Autism 
   Prevalence Trends From United States Special Education Data,
   \emph{Pediatrics} \bold{115}: e277--e282.
}
\examples{
str(ex0829)
}
\keyword{datasets}
