\name{ex1111}
\alias{ex1111}
\docType{data}
\title{Chernobyl Fallout}
\description{
  The data are are the cesium  concentrations (in Bq/kg) in soil and in 
  mushrooms at 17 wooded locations in Umbria, Central Italy, from August
  1986 to  
  November 1989. Researchers wished to investigate the cesium transfer from 
  contaminated soil to plants after the Chernobyl nuclear power plant accident 
  in April 1986 by describing the distribution of the mushroom concentration as 
  a function of soil concentration. 
}
\usage{ex1111}
\format{
  A data frame with 17 observations on the following 2 variables.
  \describe{
    \item{\code{Mushroom}}{the cesium concentration in mushrooms, Bq/kg} 
    \item{\code{Soil}}{the cesium concentration in soil, Bq/kg} 
  }
}
\source{
   Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\examples{
str(ex1111)
}
\keyword{datasets}
