\name{ex1120}
\alias{ex1120}
\docType{data}
\title{Was Tyrannosaurus Rex Warm-Blooded?}
\description{
  Data are the isotopic composition of structural bone carbonate
  (\eqn{X}) and the isotopic composition of the coexisting calcite
  cements (\eqn{Y}) in 18 bone samples from a specimen of the dinosaur
  Tyrannosaurus rex.  Evidence that the mean of \eqn{Y} is positively
  associated with \eqn{X} was used in an argument that the metabolic
  rate of this dinosaur resembled warm-blooded more than cold-blooded
  animals.
}
\usage{ex1120}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{Carbonate}}{isotopic composition of bone carbonate}
    \item{\code{Calcite}}{isotopic composition of calcite cements}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Barrick, R.E. and Showers, W.J. (1994).  Thermophysiology of
  \emph{Tyrannosaurus rex}: Evidence from Oxygen Isotopes,
  \emph{Science} \bold{265}(5169): 222--224.
}
\seealso{
  \code{\link{ex0523}}
}
\examples{
str(ex1120)
}
\keyword{datasets}
