\name{ex1122}
\alias{ex1122}
\docType{data}
\title{Deforestation and Debt}
\description{
  It has been theorized that developing countries cut down their forests
  to pay off foreign debt.  Data are debt, deforestation, and population
  from 11 Latin American nations.
}
\usage{ex1122}
\format{
  A data frame with 11 observations on the following 4 variables.
  \describe{
    \item{Country}{a character vector indicating the country}
    \item{Debt}{debt (millions of dollars)}
    \item{Deforest}{deforestation (thousands of ha)}
    \item{Pop}{population (thousands of people)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Gullison, R.R. and Losos, E.C. (1992).  The Role of Foreign Debt in
  Deforestation in Latin America, \emph{Conservation Biology}
  \bold{7}(1): 140--7.
}
\examples{
str(ex1122)
}
\keyword{datasets}
