\name{ex1514}
\alias{ex1514}
\docType{data}
\title{Melanoma and Sunspot Activity---An Observational Study}
\description{
  Several factors suggest that the incidence of melanoma is related to
  solar radiation. These data are the age-adjusted melanoma incidence among
  males in the Connecticut Tumor Registry and the sunspot activity, 1936--1972 .
}
\usage{ex1514}
\format{
  A data frame with 37 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{Melanoma}}{male melanoma incidence in number of cases
      per 100,000 population} 
    \item{\code{Sunspot}}{sunspot relative number}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Houghton, A., Munster, E.W. and Viola, M.V. (1978).  Increased
  Incidence of Malignant Melanoma After Peaks of Sunspot Activity,
  \emph{Lancet}: 759--760. 
}
\examples{
str(ex1514)
}
\keyword{datasets}
