\name{ex1515}
\alias{ex1515}
\docType{data}
\title{Lynx Trappings and Sunspots}
\description{
  Data on the annual numbers of lynx trapped in the Mackenzie River
  district of northwest Canada from 1821--1934. 
}
\usage{ex1515}
\format{
  A data frame with 114 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{Lynx}}{number of lynx trapped}
    \item{\code{Sunspots}}{number of sunspots}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Elston, C. and Nicholson, M. (1942).  The Ten Year Cycle in Numbers of
  the Lynx in Canada, \emph{Journal of Animal Ecology} \bold{11}(2):
  215--244.
}
\examples{
str(ex1515)
}
\keyword{datasets}
