\name{ex1516}
\alias{ex1516}
\docType{data}
\title{Trends in Firearm and Motor Vehicle Deaths in the U.S.}
\description{
  Data shows the number of deaths due to firearms and the number of
  deaths due to motor vehicle accidents in the United States between
  1968 and 1993.
}
\usage{ex1516}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{FirearmDeaths}}{deaths due to firearms (in thousands per
      year)} 
    \item{\code{MotorVehicleDeaths}}{deaths due to motor vehicles (in
      thousands per year)} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Data read from a Centers for Disease Control and Prevention graph
  reported in \emph{The Oregonian}, June 17, 1997. 
}
\examples{
str(ex1516)
}
\keyword{datasets}
