\name{ex1517}
\alias{ex1517}
\docType{data}
\title{S&P 500}
\description{
  The Standard and Poor's 500 stock index (S&P 500) is a benchmark of stock 
  market performance, based on 400 industrial firms, 40 financial stocks, 
  40 utilities, and 20 transportation stocks. These data include the value of a 
  $1 investment in 1871 at the end of each year from 1871 to 1999, according to 
  the S&P 500, assuming all dividends are reinvested. Describe the distribution
  of the S&P value as a function of year. 
}
\usage{ex1517}
\format{
  A data frame with 129 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{S.P500Return}}{Value of Stock at the end of the year}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\examples{
str(ex1517)
}
\keyword{datasets}
