\name{ex1519}
\alias{ex1519}
\docType{data}
\title{El Nino and the Southern Oscillation}
\description{
  The data are the Sea Surface Temperatures (SST) and Southern Oscillation
  Index (SOI) measurements from 1950 to 2010.
}
\usage{ex1519}
\format{
  A data frame with 732 observations on the following 4 variables.
  \describe{
    \item{Year}{year}
    \item{Month}{a numerical variable for month, with 1 = January}
    \item{SOI}{the Southern Oscillation Index)}
    \item{SST}{the Sea Surfact Temperature)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  U.S. National Oceanographic and Atmospheric Administration
  (\url{http://www.cpc.ncep.noaa.gov/data/indices/}).
}
\seealso{
  \code{\link{case1502}}
}
\examples{
str(ex1519)
}
\keyword{datasets}
