\name{ex1612}
\alias{ex1612}
\docType{data}
\title{Wastewater}
\description{
  Samples of effluent were divided and sent to two laboratories for
  testing.  Data are measurements of biochemical oxygen demand and
  suspended solid measurements obtained for 2 sample splits from the two
  laboratories.
}
\usage{ex1612}
\format{
  A data frame with 11 observations on the following 4 variables.
  \describe{
    \item{ComBOD}{biochemical oxygen demand measurements from
      commercial laboratory} 
    \item{ComSS}{suspended solids measurements from commercial
      laboratory} 
    \item{StaBOD}{biochemical oxygen demand measurements from
      state laboratory} 
    \item{StaSS}{suspended solids measurements from state
      laboratory} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Johnson, R.A. and Wichern, D.W. (1988).  \emph{Applied Multivariate
  Statistical Analysis}, Prentice-Hall.
}
\examples{
str(ex1612)
}
\keyword{datasets}
