\name{ex1613}
\alias{ex1613}
\docType{data}
\title{Flea Beetle Distinction}
\description{
  Data are the measurements from two very similar species of flea beetle.
}
\usage{ex1613}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{Specimen}{specimen identification number}
    \item{Jnt1}{measurement of first joint in micrometers}
    \item{Jnt2}{measurement of second joint in micrometers}
    \item{Species}{a factor with levels \code{"conc"} and
      \code{"heik"}} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Lubischew, A.A. (1962).  On the Use of Discriminant Functions in
  Taxonomy, \emph{Biometrics} \bold{18}: 455--477.
}
\examples{
str(ex1613)
}
\keyword{datasets}
