\name{ex1615}
\alias{ex1615}
\docType{data}
\title{Trends in SAT Scores}
\description{
  Data shows a partial listing of a data set with ratios of average math
  to average verbal SAT scores in the United States and the District of
  Columbia for 1989 and 1996--1999. 
}
\usage{ex1615}
\format{
  A data frame with 51 observations on the following 6 variables.
  \describe{
    \item{\code{State}}{a character vector indicating the state}
    \item{\code{M.V.89}}{average MATH SAT scores divided by average
      VERBAL SAT score in 1989} 
    \item{\code{M.V.96}}{average MATH SAT scores divided by average
      VERBAL SAT score in 1996} 
    \item{\code{M.V.97}}{average MATH SAT scores divided by average
      VERBAL SAT score in 1997} 
    \item{\code{M.V.98}}{average MATH SAT scores divided by average
      VERBAL SAT score in 1998} 
    \item{\code{M.V.99}}{average MATH SAT scores divided by average
      VERBAL SAT score in 1999} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\examples{
str(ex1615)
}
\keyword{datasets}
