\name{ex1708}
\alias{ex1708}
\docType{data}
\title{Pig Fat}
\description{
  Actual pig fat and measurements of pig fat from magnetic resonance
  images at 13 locations for 12 pigs.
}
\usage{ex1708}
\format{
  A data frame with 12 observations on the following 14 variables.
  \describe{
    \item{Fat}{actual pig fat (in percent)}
    \item{M1}{magnetic resonance image at location 1}
    \item{M2}{magnetic resonance image at location 2}
    \item{M3}{magnetic resonance image at location 3}
    \item{M4}{magnetic resonance image at location 4}
    \item{M5}{magnetic resonance image at location 5}
    \item{M6}{magnetic resonance image at location 6}
    \item{M7}{magnetic resonance image at location 7}
    \item{M8}{magnetic resonance image at location 8}
    \item{M9}{magnetic resonance image at location 9}
    \item{M10}{magnetic resonance image at location 10}
    \item{M11}{magnetic resonance image at location 11}
    \item{M12}{magnetic resonance image at location 12}
    \item{M13}{magnetic resonance image at location 13}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Glasbey, C.A and Flowler, P.A. (1992).  Regression Models Fitted Using
  Conditional Independence to Estimate Pig Fatness from Magnetic
  Resonance Images, \emph{The Statistician} \bold{41}(2): 179--184.
}
\examples{
str(ex1708)
}
\keyword{datasets}
