\name{ex2113}
\alias{ex2113}
\docType{data}
\title{Vitamin C and Colds}
\description{
  These data are from a randomized experiment to asses the effect of large doses 
  of vitamin C on the incidence of colds.
}
\usage{ex2113}
\format{
  A data frame with 4 observations on the following 4 variables.
  \describe{
    \item{Dose}{the daily dose of vitamin C, in g}
    \item{Number}{the number of subjects given that dose of vitamin C}
    \item{WithoutIllness}{the number of subjects who did not become ill}
    \item{ProportionWithout}{the proportion of subjects who did not become
      ill}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Anderson, T.W., Suranyi,G., and Beaton, G.H. (1974) The Effect on Winter 
  Illness of Large Doses of Vitamin C \emph{Canadian Medical Association Journal} 
  \bold{111} 31--36.
}
\examples{
str(ex2113)
}
\keyword{datasets}
