\name{ex2116}
\alias{ex2116}
\docType{data}
\title{Aflatoxicol and Liver Tumors in Trout}
\description{
  An experiment at the Marine/Freshwater Biomedical Sciences Center at
  Oregon State University investigated the carcinogenic effects of
  aflatoxicol, a metabolite of Aflatoxin B1, which is a toxic by-product
  produced by a mold that infects cottonseed meal, peanuts and grains.
  Twenty tanks of rainbow trout embryos were exposed to one of five
  doses of Aflatoxicol for one hour. The data represent the numbers of
  fish in each tank and the numbers of these that had liver tumours
  after one year. 
}
\usage{ex2116}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{Dose}{Dose (in ppm)}
    \item{Tumor}{Number of trout with liver tumours}
    \item{Total}{Number of trout in tank}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\examples{
str(ex2116)
}
\keyword{datasets}
