\name{ex2222}
\alias{ex2222}
\docType{data}
\title{Emulating Jane Austen's Writing Style}
\description{
  When she died in 1817, the English novelist Jane Austen had not yet
  finished the novel \emph{Sanditon}, but she did leave notes on how she
  intended to conclude the book.  The novel was completed by a ghost
  writer, who attempted to emulate Austen's style.  In 1978, a
  researcher reported counts of some words found in chapters of books
  written by Austen and in chapters written by the emulator.  These data
  are given in this data frame.
}
\usage{ex2222}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{Count}{Number of occurrences of a word in various chapters of
      books written by Jane Austen and the ghost writer}
    \item{Book}{Title of books used}
    \item{Word}{Words used}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Morton, A.Q. (1978). \emph{Literary Detection:  How to Prove
    Authorship and Fraud in Literature and Documents}, Charles
  Scribner's Sons, New York.
}
\examples{
str(ex2222)
}
\keyword{datasets}
