\name{Sleuth3Manual}
\alias{Sleuth3Manual}
\title{
  Manual of the R Sleuth3 package
}
\description{
  If the option \dQuote{pdfviewer} is set, this command will display the
  PDF version of the help pages.  
}
\usage{
Sleuth3Manual()
}
\references{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Duxbury.
}
\author{
  Berwin A Turlach \email{Berwin.Turlach@gmail.com}
}
\examples{
\dontrun{Sleuth3Manual()}
}
\keyword{documentation}
